Trouble with the controller
I spoke with Casey about this, but for some reason my `raw_stores` line
isn't working in the controller.  When I fire up a byebug or pry session
it's fine, the exact same line works.  I can even 'puts' the variable
'raw_stores' on the line above.  But when I run the test or try to use
the application, 'raw_stores' is nil.

I'm moving on to build the API for now.